Tests/LibWeb/Text/expected/ContentSecurityPolicy/style-element-nonce-idl.txt

nonce attribute: null
color: rgb(1, 2, 3)
